Simplify props, state, reducers and context for enketo forms
Our use of a react context here was uneccesary as it was rooted at
the component <FormAssignmentData/> to pass data to
<EnketoEditableForm/>, and given that it was a direct child, we
could simply use props instead. Additionally, the state that we
needed to keep track of (i.e. the xform, and the changed data)
could much more simply use useState() directly in
<EnketoEditableForm/> rather than using a reducer.