HPC-7710: include skipped questions when saving data
We have hidden the Operation field, but that's required to
determine the values for a dropdown later on. When we don't save
all data, then when we save a form, this data is lost and the
select is not correctly populated.

This change ensures we save it even though it's skipped,
following the behaviour for the enketo server.