HPC-7710: Make it clear to the user when form is validating
Enketo can cause a long delay and freeze up as part of the form
validation process. To reduce confusion, we now change the text of
the submit button to indicate when validation is happening, and
disable the ability to dismiss the validation message by clicking
on the faded background, as that could sometimes preemptively
dismiss the dialog message. Instead, users now have to press Okay.