✨ Return email in access result for active grants
Sometimes it is desireable to know the email address of users that
have access to something, or that have modified the access of others
(in addition to their names).
For example when cross-referencing an invite list of emails with the
current emails that have been granted access.

This will still only be visible for those that can see access settings
for a particular target