✨ HPC-8318: Simplify assignment state and remove canModifyWhenClean
To simplify assignment permissions, and modify the behaviour of
`editable` to indicate if the user has permission to edit the form,
we've removed `canModifyWhenClean`, and modify the UI display based
solely on assignment state and the `editable` flag.

The form state message has also been updated to reflect to the user
when inability to edit the form may be down to the reporting window
state.