Fix type error after upgrading to Typescript 4
After upgrading to Typescript 4, type union
`[Primitive, ...Primitive[]] | [{ [id: string]: Primitive }]`
cannot be used to type an array, because we get an error:
"Each member of the union type ... has signatures,
but none of those signatures are compatible with each other"

This can be resolved if for example we replace
`string[] | number[]` type with `(string | number)[]`.
However, due to how `DepsBaseType`, we cannot use that,
therefore `Array.prototype.every.call()` is used instead.