Don't remove `husky` from Docker image
If husky is removed in this script, Docker
image cannot be built.

That is because `npm run remove-unneeded-deps`
calls this script and then `npm install` is run
in Docker container.

After package installation, `prepare` lifecycle
hook is executed, which inits `husky` with `husky install`,
but if we remove `husky` this command will fail.