Create FTS Admin project
This has been created in the CDM monorepo
as it relies on the libs in this repository.

The https://github.com/UN-OCHA/hpc-apps-core repository
was originally created to house these libs and
https://github.com/UN-OCHA/hpc-ftsadmin-site repository
was created to host FTS Admin rewrite.

`hpc-apps-core` was meant to be imported in
both CDM (here) and FTS Admin (in hpc-ftsadmin-site),
but this has not been possible without major changes
to CDM and potentially removing Nx from CDM,
due to the way typescript/Nx are set up.

CDM is used in TS config of hpc_service as its
baseUrl is set to external/hpc_cdm, but that
prevents other external paths to be included.

Extracting hpc-apps-core into npm package, like it's
done for hpc-api-core may be an option in the future.