HPC-9034: Remove deprecated usages of `<Route render`
In React Router v6, the maintainers switched
away from React components to React elements.

Thus, hooks are now favored way of getting the
data, so we can no longer get route params
in methods of `<Route>`. Therefore, we need to
use `usaParams()` hook to get URL params and
parse them. In case of an error, plain JS `Error`
object is throw, so that `errorElement` defined
on the caller can define how to handle an error.