Replace `ci.sh` with plain workflow step
The idea is to remove usage of shell scripts
in CI and keep everything under the workflow file.

The shell script was just installing packages
so that it could run `lint` and `build` commands.

By reading the code of custom GitHub Action, it was
found that `ci.sh` is executed if branch isn't one of
`env/prod`, `env/stage`, `env/red.dev` or `env/blue.dev`,
but now they will always be run on affected apps.

Before, with `ci.sh`, the commands would be
run somewhere in between other stuff that
custom GitHub Action does, but now the commands
are run before anything else in custom
GitHub Action.

These two changes in behavior aren't seen
as potentially problematic.