Refactor `BrowserRouter` to `RouterProvider` for hpc-cdm
https://reactrouter.com/6.28.0/components/routes#routes
In `hpc-cdm` there are many `<Routes />` elements,
for the way is built it makes sense to let these `<Routes />`
where they are instead of briging all of them to `main.tsx`
since we have logic inside these `<Routes />`.
In documentation they say that from `v6` and on, it is an
'uncommon practice', but for the way this was thought out,
it would need a major rewrite or to let as it is for
readability and maintenence ease purpose.