Hardcode apps to build
The workflow with action `nrwl/nx-set-shas` has failed
on last attempt of production deployment, because it
tries to find the last successful production deployment
CI run, and since we haven't released in a long time,
it has failed, because the last successful run has expired.

During investigation on how to replace the calculation
of head and base revisions we can pass to `nx affected`,
it was determined that this command will always mark
all apps and libraries as affected when we bump the
version, since all of them share common `package.json`.

Nx supports individual `package.json` files per-app,
but we intentionally took singular `package.json` approach,
so it means we'll always release both apps.

Writing a dedicated script which would ignore commits
that only bump version when finding affected projects
has proved to be a complex task for the benefit it
gives, thus we'll likely drop the `apps-to-build` argument
passed to our custom GitHub Action and just always build all
images under `dockerImages` in `.github/workflow.config.json`